chore: Move eye initializer to a separate file
